Define mapInstance.
Motivation:

There is substantial flexibility in how a ServiceDiscovery
implementation chooses the definition of its Instance: so long as it's
Hashable, it's available as a choice.

This wide flexibility causes some risks in compatibility. A library that
wants to consume a generic ServiceDiscovery implementation may well want
an Instance that is not identical to the one being provided by the
implementation a specific user has.

This is a step towards allowing users to transform their
ServiceDiscovery implementations generically, by providing a way to map
instances from one type to another. This work is intentionally
restricted: we're starting with direct mapping functions that do not
block and do not require asynchony as it's easier to get right and
provides the biggest initial jump in utility. Future work may choose to
add similar functionality for mapping Service types and for supporting
asynchrony.

Modifications:

- Implemented MapInstanceServiceDiscovery.
- Defined ServiceDiscovery.mapInstance
- Wrote tests.

Results:

Users can begin to bridge the impedence mismatches between service
discovery implementations and their consumers.